chatgpt/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.3
openai==1.54.3
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
redis==5.2.0
pydantic==2.9.2
pydantic-settings==2.6.0
httpx==0.27.2
sse-starlette==2.1.3
